Fix build errors on non-windows platforms
