+++ Source/Common/Private/Memory/LockedFixedSizeAllocator.cpp
@@ -13,6 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+#include "Memory/AllocatorBase.h"
 #include "Memory/LockedFixedSizeAllocator.h"
 #include "Error.h"
 #include <algorithm>
